Fix are_in_same_session: i64 vs i32 type mismatch blocking all RTC signaling
SELECT 1 returns SQL INT4 (i32), but code expected i64 (INT8). This caused
every rtc_offer and rtc_answer to be rejected with a ColumnDecode error,
completely preventing WebRTC peer connections from forming.

+++ crates/chatalot-db/src/repos/voice_repo.rs
@@ -147,7 +147,7 @@ pub async fn are_in_same_session(
     user_a: Uuid,
     user_b: Uuid,
 ) -> Result<bool, sqlx::Error> {
-    let row: Option<(i64,)> = sqlx::query_as(
+    let row: Option<(i32,)> = sqlx::query_as(
         r#"
         SELECT 1 FROM voice_session_participants a
         JOIN voice_session_participants b ON a.session_id = b.session_id
